LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/text/2d.text.font.parse.complex.worker-expected.txt


FAIL 2d assert_equals: ctx.font === 'italic small-caps 12px "Unknown Font", sans-serif' (got small-caps italic 400 12px/2 Unknown Font, sans-serif[string], expected italic small-caps 12px "Unknown Font", sans-serif[string]) expected "italic small-caps 12px \"Unknown Font\", sans-serif" but got "small-caps italic 400 12px/2 Unknown Font, sans-serif"
